requirements.txt
asgiref==3.2.7
Django==3.0.6
django-ipware==2.1.0
pytz==2020.1
sqlparse==0.3.1
